// src/app/modules/admission/models/student-bulk-entry.ts
export class StudentBulkEntry {
    ORG_CODE: string;
    CAMPUS_CODE: string;
    CLASS_CODE: string;
    GROUP_CODE: string;
    SECTION_CODE: string;
    YEAR_CODE: string;
    SEMESTER_CODE: string;
    SHIFT_CODE: string;
    VERSION_CODE: string;
    START_SESSION_CODE: string;
    CURRENT_SESSION_CODE: string;
    ADMISSION_DATE: any;
    CLASS_ROLL: any;
    STUDENT_NAME: any;
    FATHERS_NAME: any;
    SMS_MOBILE_NUM: any;
    RowStatus: number;
    USER_NAME: string;
}